#110 - Feat : 피드 컴포넌트 수정
